Allow generating certs and assume compiler pool mode
